FPVTL-494: 'Review Additional Application' task update
