+++ src/main/java/uk/gov/hmcts/reform/prl/controllers/SendAndReplyController.java
@@ -23,8 +23,8 @@ import uk.gov.hmcts.reform.prl.models.dto.ccd.CallbackResponse;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.models.sendandreply.Message;
 import uk.gov.hmcts.reform.prl.services.EventService;
+import uk.gov.hmcts.reform.prl.services.SendAndReplyCommonService;
 import uk.gov.hmcts.reform.prl.services.SendAndReplyService;
-import uk.gov.hmcts.reform.prl.services.UploadAdditionalApplicationService;
 import uk.gov.hmcts.reform.prl.services.tab.alltabs.AllTabServiceImpl;
 import uk.gov.hmcts.reform.prl.utils.CaseUtils;
 import uk.gov.hmcts.reform.prl.utils.ElementUtils;
@@ -38,9 +38,6 @@ import java.util.stream.Collectors;
 
 import static java.util.Optional.ofNullable;
 import static org.apache.commons.collections.CollectionUtils.isEmpty;
-import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.AWP_ADDTIONAL_APPLICATION_BUNDLE;
-import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.AWP_STATUS_CLOSED;
-import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.AWP_STATUS_IN_REVIEW;
 import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.CASE_TYPE_OF_APPLICATION;
 import static uk.gov.hmcts.reform.prl.enums.sendmessages.SendOrReply.REPLY;
 import static uk.gov.hmcts.reform.prl.enums.sendmessages.SendOrReply.SEND;
@@ -58,11 +55,7 @@ public class SendAndReplyController extends AbstractCallbackController {
     private final SendAndReplyService sendAndReplyService;
     private final ElementUtils elementUtils;
     private final AllTabServiceImpl allTabService;
-    private final UploadAdditionalApplicationService uploadAdditionalApplicationService;
-
-
-
-    public static final String MESSAGES = "messages";
+    private final SendAndReplyCommonService sendAndReplyCommonService;
 
     @Autowired
     public SendAndReplyController(ObjectMapper objectMapper,
@@ -70,12 +63,12 @@ public class SendAndReplyController extends AbstractCallbackController {
                                   SendAndReplyService sendAndReplyService,
                                   ElementUtils elementUtils,
                                   AllTabServiceImpl allTabService,
-                                  UploadAdditionalApplicationService uploadAdditionalApplicationService) {
+                                  SendAndReplyCommonService sendAndReplyCommonService) {
         super(objectMapper, eventPublisher);
         this.sendAndReplyService = sendAndReplyService;
         this.elementUtils = elementUtils;
         this.allTabService = allTabService;
-        this.uploadAdditionalApplicationService = uploadAdditionalApplicationService;
+        this.sendAndReplyCommonService = sendAndReplyCommonService;
     }
 
     @PostMapping("/about-to-start")
@@ -253,62 +246,9 @@ public class SendAndReplyController extends AbstractCallbackController {
         Map<String, Object> caseDataMap = callbackRequest.getCaseDetails().getData();
 
         if (caseData.getChooseSendOrReply().equals(SEND)) {
-            caseDataMap.put(MESSAGES, sendAndReplyService.addMessage(caseData, authorisation, caseDataMap));
-            String additionalApplicationCodeSelected = sendAndReplyService.fetchAdditionalApplicationCodeIfExist(
-                caseData, SEND
-            );
-
-            if (null != additionalApplicationCodeSelected) {
-                caseDataMap.put(
-                    AWP_ADDTIONAL_APPLICATION_BUNDLE,
-                    uploadAdditionalApplicationService
-                        .updateAwpApplicationStatus(
-                            additionalApplicationCodeSelected,
-                            caseData.getAdditionalApplicationsBundle(),
-                            AWP_STATUS_IN_REVIEW
-                        )
-                );
-            }
-
-            sendAndReplyService.sendNotificationToExternalParties(
-                caseData,
-                authorisation
-            );
-
-            //send emails in case of sending to others with emails
-            sendAndReplyService.sendNotificationEmailOther(caseData);
-            //WA - clear reply field in case of SEND
-            sendAndReplyService.removeTemporaryFields(caseDataMap, "replyMessageObject");
+            sendAndReplyCommonService.sendMessages(authorisation, caseData, caseDataMap);
         } else {
-            if (YesOrNo.No.equals(caseData.getSendOrReplyMessage().getRespondToMessage())) {
-                //Reply & close
-                caseDataMap.put(MESSAGES, sendAndReplyService.closeMessage(caseData, caseDataMap));
-
-                // Update status of Additional applications if selected to Closed
-                String additionalApplicationCodeSelected = sendAndReplyService.fetchAdditionalApplicationCodeIfExist(
-                    caseData, REPLY
-                );
-                log.info("additionalApplicationCodeSelected while closing message {}", additionalApplicationCodeSelected);
-                if (null != additionalApplicationCodeSelected) {
-                    caseDataMap.put(
-                        AWP_ADDTIONAL_APPLICATION_BUNDLE,
-                        uploadAdditionalApplicationService
-                            .updateAwpApplicationStatus(
-                                additionalApplicationCodeSelected,
-                                caseData.getAdditionalApplicationsBundle(),
-                                AWP_STATUS_CLOSED
-                            )
-                    );
-                }
-
-                // in case of reply and close message, removing replymessageobject for wa
-                sendAndReplyService.removeTemporaryFields(caseDataMap, "replyMessageObject");
-            } else {
-                //Reply & append history
-                caseDataMap.put(MESSAGES, sendAndReplyService.replyAndAppendMessageHistory(caseData, authorisation, caseDataMap));
-            }
-            //WA - clear send field in case of REPLY
-            sendAndReplyService.removeTemporaryFields(caseDataMap, "sendMessageObject");
+            sendAndReplyCommonService.replyMessages(authorisation, caseData, caseDataMap);
         }
 
         //clear temp fields
